Initial commit
I had to recreate the git repository because the old test data was not
practical and it was taking up too much space in the git repo. Some new
scripts have been added: tr-integrate and tr-length. All tests are
rewritten and should be able to detect more problems than the previous
tests. The new tests did also reveal a series of bugs that have been fixed.