tr-rdf: Changed behavior of cumulative rdf in case of two groups of centers.
When one group is given, the cumulative rdf expresses the average number
of pairs up to a certain distance around one center.

When two groups are given, the cumulative rdf is the average number of
pairs a-b up to a certain distance around one center of type a. Type a
centers form the first group of atoms, while type b centers are in the
second group. Before this patch, the average was taken around centers
of type b instead of a.