tr-ic-puckering: --project feature
With the --project option, the script tr-ic-puckering will create
additional output that is useful for the analysis of spectra. It will
project the cartesian velocities of the ring atoms in the tangents of the
ring coordinates. A spectral analysis of these projected velocities will
clearly show the contribution to the spectrum of the motion along the given
coordinate. In future, all tr-ic scripts will have this feature.